feat(posts): Добавил условие для отображения ссылок редактирования и удаления поста только для автора. Ограничили доступ для НЕ авторов к соответствующим вьюшкам

+++ blog/views.py
@@ -41,6 +41,9 @@ def update_post(request, post_id):
 
     post = get_object_or_404(Post, id=post_id)
 
+    if (request.user != post.author):
+        return render(request, 'blog/not_allowed.html')
+
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
 
@@ -48,16 +51,17 @@ def update_post(request, post_id):
             form.save()
 
             return redirect("blog:post_detail", post_id=post.id)
-        else:
-            return render(request, 'blog/post_form.html', context={"form": form, 'title': title, 'submit_button_text': submit_button_text})
-
-    form = PostForm(instance=post)
+    else:
+        form = PostForm(instance=post)
 
     return render(request, 'blog/post_form.html', context={"form": form, 'title': title, 'submit_button_text': submit_button_text})
 
 
 def delete_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+
+    if (request.user != post.author):
+        return render(request, 'blog/not_allowed.html')
 
     if request.method == "POST":
         post.delete()
